image-capture/pyproject.toml
[project]
name = "image-capture"
version = "0.1.0"
description = "Image capture scripts for AgroDrone"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gpiozero>=2.0.1",
    "rpi-gpio>=0.7.1",
]
